project/requirements.txt
click
cryptography
pycryptodome
dnslib
Flask